Move switcher classes to a single namespace
